1.restructured booking_ap 2.added decorators in bookings, history views
