+++ booking_app/booking/admin.py
@@ -1,6 +1,6 @@
 from django.contrib import admin
 
-from booking_app.booking.models.model import Booking
+from booking_app.booking.model import Booking
 
 # Register your models here.
 admin.site.register(Booking)
