--- booking_app/booking/models/__init__.py
@@ -1,1 +1,0 @@
-# from .model import *
