+++ booking_app/booking/urls.py
@@ -1,9 +1,10 @@
 from django.urls import path, include
 from rest_framework.routers import DefaultRouter
-from booking_app.booking.views.views import BookingViewSet
+from booking_app.booking.views import BookingViewSet
 
 router = DefaultRouter()
 router.register(r'bookings', BookingViewSet)
+
 
 urlpatterns = [
     path('', include(router.urls)),
