+++ booking_app/history/admin.py
@@ -1,5 +1,5 @@
 from django.contrib import admin
-from booking_app.history.models.models import *
+from booking_app.history.models import *
 
 # Register your models here.
 admin.site.register(SearchHistory),
