+++ booking_app/history/urls.py
@@ -1,6 +1,7 @@
 from rest_framework.routers import DefaultRouter
 from django.urls import path, include
-from booking_app.history.views.views import *
+
+from booking_app.history.views import SearchHistoryViewSet, ViewHistoryViewSet
 
 router = DefaultRouter()
 router.register(r'search-history', SearchHistoryViewSet)
