+++ booking_app/listings/admin.py
@@ -1,6 +1,6 @@
 from django.contrib import admin
 
-from booking_app.listings.models.model import Listings
+from booking_app.listings.model import Listings
 
 # Register your models here.
 admin.site.register(Listings)
