--- booking_app/listings/models/__init__.py
@@ -1,2 +1,0 @@
-# from .model import *
-
