+++ booking_app/reviews/admin.py
@@ -1,6 +1,6 @@
 from django.contrib import admin
 
-from booking_app.reviews.models.model import Review
+from booking_app.reviews.model import Review
 
 # Register your models here.
 admin.site.register(Review)
