+++ booking_app/reviews/signals.py
@@ -1,7 +1,7 @@
 from django.db.models.signals import post_save, post_delete
 from django.dispatch import receiver
 
-from booking_app.reviews.models.model import Review
+from booking_app.reviews.model import Review
 
 
 @receiver(post_save, sender=Review)
