+++ booking_app/reviews/urls.py
@@ -1,6 +1,6 @@
 from rest_framework.routers import DefaultRouter
 from django.urls import path, include
-from booking_app.reviews.views.views import ReviewViewSet
+from booking_app.reviews.views import ReviewViewSet
 
 router = DefaultRouter()
