+++ booking_app/user/admin.py
@@ -1,6 +1,6 @@
 from django.contrib import admin
 
-from booking_app.user.models.model import Profile
+from booking_app.user.model import Profile
 
 # Register your models here.
 admin.site.register(Profile)
