+++ booking_app/user/signals.py
@@ -1,7 +1,7 @@
 from django.db.models.signals import post_save
 from django.dispatch import receiver
 from django.contrib.auth.models import User
-from booking_app.user.models.model import Profile
+from booking_app.user.model import Profile
 
 
 @receiver(post_save, sender=User)
