+++ booking_app/user/urls.py
@@ -1,20 +1,8 @@
-#
-# from booking_app.user.views.views import*
-# from django.urls import path
-#
-#
-# urlpatterns = [
-#     path('users/', UserListGenericView.as_view(), name='user-list'),
-#     path('users/register/', RegisterUserGenericView.as_view(), name='user-register'),
-#     path('users/login/', LoginUserView.as_view(), name='login_user'),
-#     path('users/logout/', LogoutUserView.as_view(), name='logout_user'),
-# ]
-#==========================================================================================
 from django.urls import path, include
 from rest_framework_simplejwt import views
 from rest_framework.routers import DefaultRouter
 
-from booking_app.user.views.views import UserViewSet, RegisterView, LoginView, LogoutView, EmailTokenObtainPairView
+from booking_app.user.views import UserViewSet, RegisterView, LoginView, LogoutView, EmailTokenObtainPairView
 
 router = DefaultRouter()
 router.register('', UserViewSet)
